// src/utils/boundaries/types.gen.ts
export type Address = {
    /**
     * Unique code of the address
     */
    code: number;
    /**
     * Feature ID of the address
     */
    feature_id: number;
    /**
     * Plot or building number of the address
     */
    plot_or_building_number: string;
    /**
     * Plot or building number of the address
     */
    building_block_number: string | null;
    /**
     * Postal code of the address
     */
    postal_code: string;
    /**
     * Street information the address belongs to
     */
    street: FlatStreet | null;
    /**
     * Residential area information the address belongs to
     */
    residential_area: FlatResidentialArea | null;
    /**
     * Municipality information the address belongs to
     */
    municipality: ShortMunicipality;
    /**
     * Point geometry of the address
     */
    geometry: Geometry;
};
// ...
export type FlatResidentialArea = {
    /**
     * Unique code of the residential area
     */
    code: number;
    /**
     * Feature ID of the residential area
     */
    feature_id: number;
    /**
     * Name of the residential area
     */
    name: string;
};

export type FlatStreet = {
    /**
     * Unique code of the street
     */
    code: number;
    /**
     * Feature ID of the street
     */
    feature_id: number;
    /**
     * Name of the street
     */
    name: string;
    /**
     * The full name of the street, including its type
     */
    full_name: string;
};

export type Geometry = {
    /**
     * Spatial Reference Identifier (SRID) for the geometry
     */
    srid: number;
    /**
     * Geometry data in WKB (Well-Known Binary) format, represented as a hex string
     */
    data: string;
};
// ...
export type Municipality = {
    /**
     * Unique code of the municipality
     */
    code: number;
    /**
     * Feature ID of the municipality
     */
    feature_id: number;
    /**
     * Name of the municipality
     */
    name: string;
    /**
     * County information the municipality belongs to
     */
    county: ShortCounty;
    /**
     * Area of the municipality in hectares
     */
    area_ha: number;
    /**
     * Date of creation of the municipality
     */
    created_at: string;
};
// ...
export type ResidentialArea = {
    /**
     * Unique code of the residential area
     */
    code: number;
    /**
     * Feature ID of the residential area
     */
    feature_id: number;
    /**
     * Name of the residential area
     */
    name: string;
    /**
     * Municipality information the residential area belongs to
     */
    municipality: ShortMunicipality;
    /**
     * Area of the residential area in hectares
     */
    area_ha: number;
    /**
     * Date of creation of the residential area
     */
    created_at: string;
};
// ...
export type Room = {
    /**
     * Unique code of the room
     */
    code: number;
    /**
     * Room number in the building or building section
     */
    room_number: string;
    /**
     * Date of creation of the room address
     */
    created_at: string;
    /**
     * Point geometry of the address
     */
    geometry: Geometry;
    /**
     * Address of the room
     */
    address: ShortAddress;
};
// ...
export type ShortAddress = {
    /**
     * Unique code of the address
     */
    code: number;
    /**
     * Feature ID of the address
     */
    feature_id: number;
    /**
     * Plot or building number of the address
     */
    plot_or_building_number: string;
    /**
     * Plot or building number of the address
     */
    building_block_number: string | null;
    /**
     * Postal code of the address
     */
    postal_code: string;
    /**
     * Street information the address belongs to
     */
    street: FlatStreet | null;
    /**
     * Residential area information the address belongs to
     */
    residential_area: FlatResidentialArea | null;
    /**
     * Municipality information the address belongs to
     */
    municipality: ShortMunicipality;
};

export type ShortCounty = {
    /**
     * Unique code of the county
     */
    code: number;
    /**
     * Feature ID of the county
     */
    feature_id: number;
    /**
     * Name of the county
     */
    name: string;
};

export type ShortMunicipality = {
    /**
     * Unique code of the municipality
     */
    code: number;
    /**
     * Feature ID of the municipality
     */
    feature_id: number;
    /**
     * Name of the municipality
     */
    name: string;
    /**
     * County information the municipality belongs to
     */
    county: ShortCounty;
};

export type ShortResidentialArea = {
    /**
     * Unique code of the residential area
     */
    code: number;
    /**
     * Feature ID of the residential area
     */
    feature_id: number;
    /**
     * Name of the residential area
     */
    name: string;
    /**
     * Municipality information the residential area belongs to
     */
    municipality: ShortMunicipality;
};

export type Street = {
    /**
     * Unique code of the street
     */
    code: number;
    /**
     * Feature ID of the street
     */
    feature_id: number;
    /**
     * Name of the street
     */
    name: string;
    /**
     * The full name of the street, including its type
     */
    full_name: string;
    /**
     * The total length of the street in meters
     */
    length_m: number;
    /**
     * Date of creation of the street
     */
    created_at: string;
    /**
     * Residential area information the street belongs to
     */
    residential_area: ShortResidentialArea;
};